+++ src/reflect/_MulleObjC-versioncheck.h
@@ -7,7 +7,7 @@
 
 #if defined( MULLE__OBJC__DEBUG_VERSION)
 # ifndef MULLE__OBJC__DEBUG_VERSION_MIN
-#  define MULLE__OBJC__DEBUG_VERSION_MIN  ((0UL << 20) | (21 << 8) | 2)
+#  define MULLE__OBJC__DEBUG_VERSION_MIN  ((0UL << 20) | (21 << 8) | 3)
 # endif
 # ifndef MULLE__OBJC__DEBUG_VERSION_MAX
 #  define MULLE__OBJC__DEBUG_VERSION_MAX  ((0UL << 20) | (22 << 8) | 0)
@@ -22,10 +22,10 @@
 
 #if defined( MULLE__OBJC__RUNTIME_VERSION)
 # ifndef MULLE__OBJC__RUNTIME_VERSION_MIN
-#  define MULLE__OBJC__RUNTIME_VERSION_MIN  ((0UL << 20) | (24 << 8) | 0)
+#  define MULLE__OBJC__RUNTIME_VERSION_MIN  ((0UL << 20) | (25 << 8) | 0)
 # endif
 # ifndef MULLE__OBJC__RUNTIME_VERSION_MAX
-#  define MULLE__OBJC__RUNTIME_VERSION_MAX  ((0UL << 20) | (25 << 8) | 0)
+#  define MULLE__OBJC__RUNTIME_VERSION_MAX  ((0UL << 20) | (26 << 8) | 0)
 # endif
 # if MULLE__OBJC__RUNTIME_VERSION < MULLE__OBJC__RUNTIME_VERSION_MIN
 #  error "mulle-objc-runtime is too old"